+++ packages/typespec-rust/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed formatting of some doc comments.
   * HTML elements are converted to markdown equivalents.
   * Bare URLs are converted to Rust docs hyperlinks.
+* The emitter will attempt to execute `cargo fmt` after files are written.
 
 ## 0.8.2 (2025-02-04)